pyproject.toml
[project]
name = "dialogue-generation"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-speechmatics~=1.2",
    "python-dotenv>=1.2.1",
]
